moved category files in one folder
