+++ Tabloid/client/src/components/ApplicationViews.js
@@ -5,8 +5,8 @@ import Login from "./Login";
 import Register from "./Register";
 import Hello from "./Hello";
 import UserList from "./UserProfileList";
-import CategoryList from "./CategoryList";
-import CategoryForm from "./CategoryForm";
+import CategoryList from "./Category/CategoryList";
+import CategoryForm from "./Category/CategoryForm";
 import TagList from "./TagList";
 import TagForm from "./TagForm";
