+++ Tabloid/client/src/components/Category/Category.js
